Fixed error in SimpleMultiplexer where an array was indexed with Integer instead of int.

+++ src/com/stefankopieczek/audinance/conversion/multiplexers/SimpleMultiplexer.java
@@ -17,7 +17,8 @@ public class SimpleMultiplexer implements Multiplexer
 	                                Integer targetNumChannels)
 	{
 		DecodedSource[] oldChannels = result.getChannels();
-		DecodedSource[] newChannels = new DecodedSource[targetNumChannels];
+		DecodedSource[] newChannels = 
+	  	              new DecodedSource[targetNumChannels.intValue()];
 		
 		if (targetNumChannels > oldChannels.length)
 		{
